gmaps_pipeline/requirements.txt
requests>=2.28
numpy>=1.24
Pillow>=9.5
python-dotenv>=1.0
